pyproject.toml
[tool.poetry]
name = "portfolio_backend"
version = "0.0.1"
description = ""
authors = ["Parametricall <22248105+Parametricall@users.noreply.github.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "3.9.1"
django = "^3.1.1"
djangorestframework = "^3.12.2"
django-cors-headers = "^3.6.0"
drf-writable-nested = "^0.6.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
